Add a fake dependency on the konsole4-kpart
This fake dependency is needed to guarantee upgrades from jessie to stretch,
at least while we are still supporting kde4 packages. The kde4 version of
konsole provided the kpart in the bin:konsole, and users of the kpart (like
yakuake) depend on this in order to work. We would change these packages to
depend on konsole4-kpart. When we pull the plug to the kde4 packages we would
need a breaks kde4libs somewhere to ensure smooth upgrades afterwards.